% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchEffects.R
\name{addBatchCorrection}
\alias{addBatchCorrection}
\title{Add batch correction assays and metrics}
\usage{
addBatchCorrection(
  exp,
  assay = "ratio",
  qcType = metadata(exp)$QC,
  removeOutliers = TRUE,
  useWithinBatch = TRUE,
  calculateRSDQC = FALSE
)
}
\arguments{
\item{exp}{SummarizedExperiment Object}

\item{assay}{Assay to be used for batch correction}

\item{qcType}{Type of QC to be used for batch correction}

\item{removeOutliers}{Remove outliers from the batch correction}

\item{useWithinBatch}{Use within batch correction}

\item{calculateRSDQC}{Boolean value. Should the assay be used to
calculate RSDQC values? Default is FALSE.}
}
\value{
SummarizedExperiment with batch-effect corrected RSDQCs in the
rowData slot
}
\description{
Batch effects can cause the variation of A/IS ratios to be
very high. By using pooled QC samples, batch effect factors per compound
can be calculated. These are multiplied by the original A/IS ratio to obtain
batch-corrected ratios. This usually results in improved results.
}
\examples{
# Read the example dataset
exp <- readRDS(system.file("extdata/data.RDS", package = "mzQuality"))

# Add batch correctied assay and batch-corrected RSDQCs
exp <- addBatchCorrection(exp)

# Show RSDQCs
rowData(exp)$rsdqc

# Show batch-corrected RSDQCs
rowData(exp)$rsdqcCorrected
}
