% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchEffects.R
\name{addBatchCorrectionAssay}
\alias{addBatchCorrectionAssay}
\title{Calculate the batch effection corrections}
\usage{
addBatchCorrectionAssay(
  exp,
  assayName = "ratio",
  saveAssay = "ratio_corrected",
  qc = metadata(exp)$QC,
  removeOutliers = TRUE
)
}
\arguments{
\item{exp}{SummarizedExperiment Object}
}
\value{
SummarizedExperiment with batch-effect corrected ratios in the
assay slot with name "ratio Corrected.
}
\description{
Calculate the batch effection corrections
}
\examples{
# Read the example dataset
exp <- readRDS(system.file("data.RDS", package = "mzQuality"))

# Add batch correction assays
exp <- addBatchCorrectionAssay(exp)

# Show results of the assay
assay(exp, "ratio_corrected")
}
