% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{addConcentrations}
\alias{addConcentrations}
\title{Add Known Concentrations of Calibration Lines}
\usage{
addConcentrations(exp, df)
}
\arguments{
\item{exp}{A SummarizedExperiment object to which the concentration data
will be added.}

\item{df}{A data frame containing the concentrations of calibration lines.
Rows should correspond to calibration line names, and columns should
correspond to compound names.}
}
\value{
A SummarizedExperiment object with an added assay called
    "Concentration". The assay is filled with the values provided in `df`,
    and other values are set to `NA`.
}
\description{
Adds known concentrations of calibration lines (CAL) to a
    SummarizedExperiment object. This function creates a new assay called
    "Concentration" in the experiment, which is filled with `NA` values
    and the supplied concentrations.
}
\details{
The concentrations should be provided as a data frame where rows
    represent the names of calibration lines and columns represent
    compound names. Optionally, the `filterComps` argument can be set to
    TRUE to retain only compounds in the experiment that have a known
    concentration in the CAL lines.
}
\examples{
# Read example dataset
exp <- readRDS(system.file("extdata/data.RDS", package = "mzQuality"))
# Add concentrations
file <- system.file("extdata/concentrations.tsv", package = "mzQuality")
concentrations <- read.delim(file)

addConcentrations(exp, df)
}
