% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{addLabels}
\alias{addLabels}
\title{Add labels to a plot}
\usage{
addLabels(p, textColor = "white", textSize = 3)
}
\arguments{
\item{p}{A ggplot object to which labels will be added.}

\item{textColor}{A character string specifying the color of the text.
Defaults to "white".}

\item{textSize}{A numeric value specifying the size of the text.
Defaults to 3.}
}
\value{
A ggplot object with added labels.
}
\description{
Adds labels to a ggplot object using ggrepel.
}
\details{
This function uses ggrepel to add labels to a ggplot object. It
    requires the ggrepel package to be installed.
}
