% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calRange.R
\name{addLinearRange}
\alias{addLinearRange}
\title{Add the ratio of compounds present in linear range}
\usage{
addLinearRange(
  exp,
  assay = "ratio",
  calType = metadata(exp)$concentration,
  minCalNo = 2,
  maxCalNo = 6,
  saveAssay = "CalRange"
)
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
