% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{addPlotConfidenceInterval}
\alias{addPlotConfidenceInterval}
\title{Adding confidence lines}
\usage{
addPlotConfidenceInterval(plot, df, assay)
}
\arguments{
\item{plot}{A ggplot object to which confidence interval lines will be
added.}

\item{df}{A data frame containing the data used to calculate confidence
intervals. Must include columns for `Aliquot` and the assay.}

\item{assay}{A character string specifying the column name of the assay
to be used for calculating confidence intervals.}
}
\value{
A ggplot object with added confidence interval lines.
}
\description{
Adds confidence interval lines to a ggplot object.
}
\details{
This function calculates the mean and standard deviation of
    medians for each aliquot and adds dashed horizontal lines to the plot
    at ±3 standard deviations from the mean.
}
