% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{aliquotPlot}
\alias{aliquotPlot}
\title{Aliquot Plot}
\usage{
aliquotPlot(
  exp,
  assay,
  batches = exp$batch,
  types = exp$type,
  logTransform = FALSE
)
}
\arguments{
\item{exp}{A data frame containing experimental data. Must include columns
for `batch`, `type`, `Aliquot`, and `injection_time`.}

\item{assay}{A character string specifying the column name of the assay
to be plotted on the y-axis.}

\item{batches}{A vector specifying the batches to include in the plot.
Defaults to `exp$batch`.}

\item{types}{A vector specifying the types to include in the plot.
Defaults to `exp$type`.}

\item{logTransform}{A logical value indicating whether to apply log scaling
to the y-axis. Defaults to FALSE.}
}
\value{
A ggplot object representing the aliquot plot.
}
\description{
Creates a boxplot for aliquots with optional log scaling.
}
\details{
This function generates a boxplot for aliquots using ggplot2. It
    allows for optional log scaling of the y-axis and customizes the plot
    based on the provided experimental data.
}
\examples{
# Read data
path <- system.file("extdata", "example.tsv", package = "mzQuality")
exp <- buildExperiment(readData(path))

# Do Analysis
exp <- doAnalysis(exp)

# Create an aliquot plot
aliquotPlot(
   exp,
   assay = "ratio_corrected",
   types = c("SQC", "LQC"),
   logTransform = TRUE
)
}
