% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundPlotNew.R
\name{aliquotPlotNew}
\alias{aliquotPlotNew}
\title{New Aliquot Plot}
\usage{
aliquotPlotNew(
  exp,
  assay,
  batches = exp$batch,
  types = exp$type,
  doLog = FALSE
)
}
\value{
placeholder
}
\description{
A short description...
}
\details{
placeholder
}
