% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlankEffects.R
\name{backgroundSignals}
\alias{backgroundSignals}
\title{Calculate background signals in compounds}
\usage{
backgroundSignals(exp, type = "BLANK", NaAsZero = FALSE)
}
\arguments{
\item{exp}{SummarizedExperiment Object}

\item{type}{Name of the sample type representing blank samples.}

\item{NaAsZero}{Logical, if TRUE, all NA values are replaced with 0 before
calculating the background signal.}
}
\value{
SummarizedExperiment with an added "backgroundSignal" column in
the rowData slot
}
\description{
Background signal (also called blank effect) is the relative
intensity measured in blank samples compared to study samples. Here, the
background signal is calculated for the sample with the given blank type.
This is done by dividing the means blank signal by the median sample signal.
If this value exceeds 0.4, it is generally cosidered having a poor
signal-to-noise (SNR) ratio. Consider removing compounds with a poor SNR.
}
\examples{
# Read the example dataset
exp <- readRDS(system.file("extdata/data.RDS", package = "mzQuality"))

# Perform the calculation
backgroundSignals(exp)
}
