% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{batchCorrectionPlot}
\alias{batchCorrectionPlot}
\title{Plot the batch correction factors}
\usage{
batchCorrectionPlot(exp)
}
\arguments{
\item{exp}{SummarizedExperiment Object}
}
\value{
ggplot object of a boxplot of batch-effect correction factors
}
\description{
The batch correction plot shows a boxplot of correction factors
per batch. These are the calculated multipliers, which ideally should be
as close to 1. A value near 1 indicates the batch is similar to the median
of all batches. This plot can be used to see if compounds are
over-corrected, or if a batch has bad measurements.
}
\examples{
# Read the example dataset
exp <- readRDS(system.file("data.RDS", package = "mzQuality"))

# Create boxplot with batch-correction values
batchCorrectionPlot(exp)
}
