% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlankEffects.R
\name{blankLimits}
\alias{blankLimits}
\title{Add limits of Blank, Detection and Quantification with Concentrations}
\usage{
blankLimits(exp, type = "PROC", assay = "concentration")
}
\arguments{
\item{exp}{A SummarizedExperiment object}

\item{type}{Name of the sample type representing processed blanks,
defaults to 'PROC'}
}
\value{
SummarizedExperiment with 3 added columns in the rowData slot
called "LoB" (Limit of Blank), "LoD" (Limit of Detection), and "LoQ"
(Limit of Quantification)
}
\description{
The Limit of Blank (LoB), -Detection (LoD) and -Quantification
(LoQ) are metrics used to determine if the calculated absolute
quantification is reliable. Compounds with calculated concentrations that
approach these limits might be unreliable. The LoB represent the 95%
percentile of the Blank samples, the LoD represents the limit where a
compound can reliably be detected. The LoQ represents the limit where a
compound can reliably be quantified. The results of the LoB, LoD, and LoQ
are stored in the rowData of the experiment.
}
\examples{
# Read example dataset
data <- read.delim(system.file(package = "mzQuality", "dataset.txt"))

# Construct experiment
exp <- buildExperiment(
    data,
    rowIndex = "Compound",
    colIndex = "Aliquot",
    primaryAssay = "Area",
    secondaryAssay = "Area_is"
)

# Add concentrations
concentrations <- read.delim(system.file(
    package = "mzQuality",
    "concentrations.txt"
))
exp <- addConcentrations(exp, concentrations)

# Perform analysis
exp <- doAnalysis(exp)

# Model concentrations
exp <- calculateConcentrations(exp)

# Calculate limits of blank samples
exp <- blankLimits(exp)
exp
}
