% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{buildMetadataWithIndex}
\alias{buildMetadataWithIndex}
\title{Build metadata DataFrame using an index column}
\usage{
buildMetadataWithIndex(df, index)
}
\arguments{
\item{df}{A dataframe in long format containing the data.}

\item{index}{A string specifying the column to use as the index.}
}
\value{
A DataFrame containing the metadata, ordered by the unique values
    in the index column.
}
\description{
Constructs a metadata DataFrame from a long-format dataframe
    using the specified index column.
}
\details{
This function extracts metadata columns from the input dataframe
    using `getMetaData()`, then formats and orders the metadata using
    `setMetaDataFrame()`.
}
