% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalLinesHighPerformance.R
\name{calculateConcentrations}
\alias{calculateConcentrations}
\title{Calculate concentratios for a full experiment}
\usage{
calculateConcentrations(exp, byBatch = TRUE, saveAssay = "concentration", ...)
}
\arguments{
\item{byBatch}{Should the concentration be calculated per batch? Defaults to
TRUE.}

\item{saveAssay}{Name of the assay to save the results. Defaults to
`"Concentration"`}

\item{...}{
  Arguments passed on to \code{\link[=batchConcentrations]{batchConcentrations}}
  \describe{
    \item{\code{exp}}{SummarizedExperiment object}
    \item{\code{checkForOutliers}}{Boolean value, should studentized residuals be
used to determine outliers? If TRUE, outliers are detected and only removed
at the start or end of a calibration line. If FALSE, outlier detection
is ignored.}
    \item{\code{useWeights}}{Boolean value, should weighted regression be used? Defaults
to FALSE, which indicates no weights are used.}
  }}
}
\value{
SummarizedExperiment with predicted concentrations in the assay
given in `saveAssay`. Defaults to `"concentration"`.
}
\description{
Calculate concentratios for a full experiment
}
