% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSDQCs.R
\name{calculateCorrectedRSDQCs2}
\alias{calculateCorrectedRSDQCs2}
\title{Calculate the RSDQC of an assay}
\usage{
calculateCorrectedRSDQCs2(
  exp,
  primaryAssay = "area",
  secondaryAssay = "area_is",
  qcType = "SQC",
  returnRSDs = FALSE,
  saveAssay = "ratio_corrected"
)
}
\arguments{
\item{exp}{SummarizedExperiment object}
}
\description{
The RSDQC is an indicator of variance of compounds in QC
samples. This function returns the RSDQC of all compounds in the
SummarizedExperiment
}
\details{
placeholder
}
\examples{
# Read the example dataset
exp <- readRDS(system.file("data.RDS", package = "mzQuality"))

calculateCorrectedRSDQCs2(exp)
}
