% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{calculateRatio}
\alias{calculateRatio}
\title{Calculate the A/IS ratio with minimal imputation}
\usage{
calculateRatio(
  exp,
  assay1 = metadata(exp)$primary,
  assay2 = metadata(exp)$secondary
)
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{assay1}{Name of the primary assay, usually "Area" for compounds}

\item{assay2}{Name of the secondary assay, usually "Area_is" for internal
standards}
}
\value{
SummarizedExperiment with an assay slot called "ratio" containing
the ratio between two assays.
}
\description{
Calculate the A/IS ratio with minimal imputation
}
\examples{
# Read example dataset
data <- read.delim(system.file(package = "mzQuality", "dataset.txt"))

# Construct experiment
exp <- buildExperiment(
    data,
    rowIndex = "Compound",
    colIndex = "Aliquot",
    primaryAssay = "Area",
    secondaryAssay = "Area_is"
)

# Calculate the ratio
exp <- calculateRatio(exp)

# Show results
assay(exp, "ratio")
}
