% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{calibrationPlot}
\alias{calibrationPlot}
\title{Create calibration plot}
\usage{
calibrationPlot(
  exp,
  assay = "ratio_corrected",
  compound = 1,
  batch = 1,
  guides = NULL
)
}
\arguments{
\item{exp}{SummarizeExperiment object}

\item{assay}{which assay to plot}

\item{compound}{Which compound to plot}

\item{batch}{Batch number}

\item{guides}{Which guides to add. Options available are c("95% CI",
"Limit of Blank", "Limit of Detection", "Limit of Quantification")}
}
\value{
ggplot object of a scatterplot with "SAMPLE", "QC" and "CAL" types.
}
\description{
The calibration plot shows the CAL Lines, Samples, and QCs of
a single compound in a different view to the compoundPlot. The CAL lines
are here shown as lines, which aids in determining if the samples are
between the lowest and highest CAL line.
}
\examples{
# Read the example dataset
exp <- readRDS(system.file("data.RDS", package = "mzQuality2"))

# Create scatterplot with calibration lines
calibrationPlot(exp)
}
