% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlankEffects.R
\name{carryOverEffect}
\alias{carryOverEffect}
\title{Calculate Carry-over Effect}
\usage{
carryOverEffect(exp, type = "PROC", assay = "concentration")
}
\arguments{
\item{exp}{A SummarizedExperiment object}

\item{type}{Name of the sample type that represents a blank sample,
defaults to 'BLANK'.}

\item{assay}{Name of the assay to use for the calculation, defaults to
'concentration'.}
}
\value{
SummarizedExperiment with an added column in the rowData slot
called "CarryOver"
}
\description{
The carry-over effect represents a percentage of the compound
concentration that has been measured in a blank sample following another
sample. Ideally this value should be as low as possible. In order to run
this function, the calculations for all samples must be calculated using
`calculateConcentrations()`.
}
\examples{
# Read the example dataset
exp <- readRDS(system.file("extdata/data.RDS", package = "mzQuality"))

# Add concentrations
concentrations <- read.delim(system.file(
    package = "mzQuality",
    "extdata/concentrations.tsv"
))
exp <- addConcentrations(exp, concentrations)

# Perform analysis
exp <- doAnalysis(exp, doAll = TRUE)

# Model concentrations
exp <- calculateConcentrations(exp)

# Calculate carryOverEffect
exp <- carryOverEffect(exp)
exp
}
