% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{cleanDataframe}
\alias{cleanDataframe}
\title{Clean a Dataframe by Removing Non-Numeric or Matrix Columns}
\usage{
cleanDataframe(df, onlyNumeric = FALSE)
}
\arguments{
\item{df}{A dataframe to be cleaned.}

\item{onlyNumeric}{Logical. If TRUE, retains only numeric columns in the
dataframe. Defaults to FALSE.}
}
\value{
A cleaned dataframe with the specified columns removed.
}
\description{
Cleans a dataframe by removing columns of class "matrix" and,
    optionally, retaining only numeric columns.
}
\examples{
# Example dataframe
df <- data.frame(
    A = matrix(1:4, nrow = 2),
    B = c(1.5, 2.5),
    C = c("x", "y")
)

# Remove matrix columns
cleanDataframe(df)

# Remove matrix columns and retain only numeric columns
cleanDataframe(df, onlyNumeric = TRUE)
}
