% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combat.R
\name{combat}
\alias{combat}
\title{Combat batch correction}
\usage{
combat(data, batch)
}
\arguments{
\item{data}{matrix of compounds x samples}

\item{batch}{Vector of batch labels, same length as columns in data}
}
\description{
A Bayes Theorem based batch correction
}
\details{
Workflow:
1. Standardize the data, so that the values (compounds) become relative
across samples. By standardizing, you're subtracting the mean and divide
by the standard deviation
2. Using the batch labels, multiply the standardized score with the batch
occurence so that a matrix of batch x compound contains the weight of each
compound for a batch (gamma_hat).
3. Calculate the mean of each batch, so we get the average values of each
batch with the standardized compounds (gamma_bar).
4. Calculate the variance per batch (gamma_var)
5. Multiply the batch variance (gamma_var) by the weights (gamma_hat) and
add the mean to get the batch-corrected weights. Then divide by the
variance + 1
6. For each compound, multiply by the batch occurence. These are the standardized
batch effects. Subtract this from the original standardized values to obtain
batch-corrected standardized values.
7. Reverse the standardization by multiplying and adding the original
standard deviation and means
8. Return the batch-corrected values
}
\examples{
# Simulate some data
set.seed(1)
n <- 100 # samples
p <- 200 #
x <- seq(10, 250, length.out = n)
batch <- sample(1:5, n, replace = TRUE)
data <- sapply(1:p, function(i) 2 * x + rnorm(n, 0, 10) + (batch - 1) * 20)
# Transpose data to have samples in columns and features in rows
data <- t(data)
corrected_data <- combat(data, batch)
# Perform PCA
pca_original <- prcomp(t(data), scale. = TRUE)
pca_corrected <- prcomp(t(corrected_data), scale. = TRUE)
# Plot PCA results for original data
par(mfrow = c(1, 2))
plot(pca_original$x[, 1], pca_original$x[, 2], col = batch, pch = 19,
     main = "PCA of Original Data", xlab = "PC1", ylab = "PC2")
# Plot PCA results for batch-corrected data
plot(pca_corrected$x[, 1], pca_corrected$x[, 2], col = batch, pch = 19,
     main = "PCA of Batch-Corrected Data", xlab = "PC1", ylab = "PC2")
}
