% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundPlotNew.R
\name{compoundPlot}
\alias{compoundPlot}
\title{Compound plot function}
\usage{
compoundPlot(df, assay, trendTypes, colors, smooth = TRUE, doLog = TRUE)
}
\value{
placeholder
}
\description{
A short description...
}
\details{
placeholder
}
