% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundPlotNew.R
\name{compoundPlotNew}
\alias{compoundPlotNew}
\title{New Compound Plot}
\usage{
compoundPlotNew(
  exp,
  compound = 1,
  batches = exp$batch,
  types = exp$type,
  assay = "ratio_corrected",
  doLog = TRUE,
  withinTrend = FALSE,
  shortName = FALSE,
  trendTypes = metadata(exp)$QC,
  addInternalStandards = FALSE,
  smooth = TRUE
)
}
\value{
placeholder
}
\description{
A short description...
}
\details{
placeholder
}
