% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{compoundReports}
\alias{compoundReports}
\title{Create dynamic compound report(s)}
\usage{
compoundReports(
  exp,
  folder,
  assays = c("ratio", "ratio_corrected"),
  project = "mzQuality"
)
}
\arguments{
\item{exp}{SummarizedExperiment object containing the experiment data.}

\item{folder}{Output folder to store the compound reports.}

\item{assays}{Character vector of assay names to visualize in the reports.
Must exist in `exp`.}

\item{project}{Character string denoting the project name. Defaults to
"mzQuality".}
}
\value{
One HTML report per compound is saved in the specified folder.
}
\description{
Generates interactive HTML reports for individual compounds
    from a SummarizedExperiment object. Each report includes compound-
    specific plots across selected assays for quality control and visual
    inspection.
}
\details{
The function parallelizes report generation over the number of
    compounds using `multiApply()`, supporting multicore environments. Each
    compound report is stored as a separate HTML file named after the
    compound.
}
\examples{
# exp <- readRDS(system.file(package = "mzQuality", "data.RDS"))
# compoundReports(exp[1:2, ], folder = tempdir())
}
