% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{concentrationPlot}
\alias{concentrationPlot}
\title{Concentration Plot}
\usage{
concentrationPlot(
  exp,
  assay = "ratio_corrected",
  compound = 1,
  batch = 1,
  types = c("SAMPLE", "SQC"),
  calType = metadata(exp)$concentration,
  removeOutliers = TRUE,
  withinTrend = TRUE,
  plotOnCalibrationLine = TRUE
)
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
