% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundPlotNew.R
\name{concentrationPlotNew}
\alias{concentrationPlotNew}
\title{New Concentration Plot}
\usage{
concentrationPlotNew(
  exp,
  assay = "ratio_corrected",
  compound = 1,
  batch = 1,
  types = c("SAMPLE", "SQC"),
  calType = metadata(exp)$concentration,
  removeOutliers = TRUE,
  withinTrend = TRUE,
  plotOnCalibrationLine = TRUE
)
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
