% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{convertExperiment}
\alias{convertExperiment}
\title{Convert an experiment to one capable with mzQuality}
\usage{
convertExperiment(
  exp,
  primaryAssay = "area",
  secondaryAssay = primaryAssay,
  istd = "compound_is",
  type = "type",
  datetime = "datetime",
  batch = "batch",
  qcType = "SQC"
)
}
\arguments{
\item{exp}{A SummarizedExperiment object}

\item{primaryAssay}{Assay that should be used for compound Areas / Intensity}

\item{secondaryAssay}{Assay that should be used for internal standard
Areas / Intensity. Defaults to equal to primaryAssay, meaning no internal
standards are present.}

\item{istd}{Column in the rowData that represents the internal standard used
for the compound. Defaults to 'compound_is'}

\item{type}{Column in the colData that represents the sample type.
Defaults to 'Type'}

\item{datetime}{Column in the colData that represents the datetime when the
sample was measured. Defaults to 'Datetime'}

\item{batch}{Column in the colData that represents the batch name / number.
Defaults to 'Batch'}

\item{qcType}{Name of the QC sample type used. Defaults to 'SQC'}
}
\value{
SummarizedExperiment that is suitable for use with mzQuality
}
\description{
This function can be used to convert an existing
SummarizedExperiment to one capable to use with mzQuality. This function
will change some colData and rowData columns, but will leave assays
untouched.
}
\details{
This function will convert the column names of the colData and
rowData to the ones used in mzQuality. The function will also add
metadata to the SummarizedExperiment object, which is used to store
information about the experiment. The metadata will be added to the
SummarizedExperiment object if it is not already present. The function
will also check if the primaryAssay and secondaryAssay are the same.
}
\examples{
# Read example dataset
data <- readData(system.file(package = "mzQuality", "example.tsv"))

# Construct experiment
exp <- buildExperiment(
    data,
    rowIndex = "compound",
    colIndex = "aliquot",
    primaryAssay = "area",
    secondaryAssay = "area_is"
)
convertExperiment(exp, primaryAssay = "Area")
}
