% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{createPcaPlot}
\alias{createPcaPlot}
\title{Create a PCA plot with optional confidence intervals}
\usage{
createPcaPlot(data, xAxis, yAxis, addConfidenceInterval = TRUE)
}
\arguments{
\item{data}{A data frame containing the PCA results. It must include
columns `PC1`, `PC2`, `type`, and `text`.}

\item{xAxis}{A string specifying the label for the x-axis.}

\item{yAxis}{A string specifying the label for the y-axis.}

\item{addConfidenceInterval}{Logical, whether to add confidence intervals
(ellipses) for each group. Defaults to `TRUE`.}
}
\value{
A ggplot object representing the PCA plot.

A ggplot object representing the PCA plot.
}
\description{
Generates a PCA plot using ggplot2, with options to add
    confidence intervals for each group.
}
\details{
This function creates a PCA plot from the provided data, with
    points colored by group and optional confidence intervals (ellipses)
    for each group. The x-axis and y-axis labels can be customized, and
    colors for the groups are specified using a color vector.
}
