% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{createReportFolder}
\alias{createReportFolder}
\title{Create a project-specific report folder structure}
\usage{
createReportFolder(baseFolder, projectName)
}
\arguments{
\item{baseFolder}{A string specifying the base directory where the project
folder will be created.}

\item{projectName}{A string specifying the name of the project. This will
be used as the name of the main project folder.}
}
\value{
None. The function creates the folder structure as a side effect.
}
\description{
Creates a folder structure for storing reports, plots, and
    compound-specific data for a given project.
}
\details{
This function generates a folder structure within a specified base
    folder for a given project. The structure includes:
    - A main project folder
    - A subfolder for exports
    - A subfolder for plots
    - A subfolder for compound-specific plots

    If the folders do not already exist, they are created recursively.
}
