% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{createReports}
\alias{createReports}
\title{Generate project reports and export data}
\usage{
createReports(
  folder,
  project,
  exp,
  summaryPlots = c("Aliquot", "PCA", "QC"),
  makeSummaryReport = TRUE,
  makeCompoundReport = TRUE,
  backgroundPercent = 40,
  cautionRSD = 15,
  nonReportableRSD = 30,
  assays = c("ratio", "ratio_corrected")
)
}
\arguments{
\item{folder}{A string specifying the base directory where the reports will
be saved.}

\item{project}{A string specifying the name of the project. This will be
used to create a project-specific folder.}

\item{exp}{A SummarizedExperiment object containing the experimental data.}

\item{summaryPlots}{A character vector specifying the types of summary plots
to include in the summary report. Defaults to
`c("Aliquot", "PCA", "QC")`.}

\item{makeSummaryReport}{Logical, whether to generate a summary report.
Defaults to `TRUE`.}

\item{makeCompoundReport}{Logical, whether to generate compound-specific
reports. Defaults to `TRUE`.}

\item{backgroundPercent}{Numeric, the percentage threshold for background
correction. Defaults to `40`.}

\item{cautionRSD}{Numeric, the RSD threshold for cautionary reporting.
Defaults to `15`.}

\item{nonReportableRSD}{Numeric, the RSD threshold for non-reportable data.
Defaults to `30`.}

\item{assays}{A character vector specifying the assay names to include in
the reports. Defaults to `c("ratio", "ratio_corrected")`.}
}
\value{
None. The function creates reports and exports data as a side effect.
}
\description{
Creates a folder structure, exports data tables, and generates
    summary and compound-specific reports for a given project.
}
\details{
This function automates the process of creating reports for a
    project. It performs the following tasks:
    - Creates a folder structure for storing reports and plots.
    - Exports data tables to the specified folder.
    - Generates a summary report with specified plots.
    - Generates compound-specific reports.

    The function allows customization of the report content and thresholds
    for background percentage and relative standard deviation (RSD).
}
\examples{
exp <- readRDS(system.file(package = "mzQuality", "data.RDS"))
createReports(
    folder = tempdir(),
    project = "MyProject",
    exp = exp[1:5, ],
    summaryPlots = c("Aliquot", "PCA"),
    makeSummaryReport = TRUE,
    makeCompoundReport = TRUE,
    backgroundPercent = 50,
    cautionRSD = 10,
    nonReportableRSD = 25,
    assays = "ratio"
)
}
