% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{doAnalysis}
\alias{doAnalysis}
\title{Perform the main analysis of mzQuality}
\usage{
doAnalysis(
  exp,
  aliquots = colnames(exp),
  doAll = FALSE,
  removeBadCompounds = FALSE,
  removeOutliers = FALSE,
  useWithinBatch = FALSE,
  effectNaAsZero = FALSE,
  concentrationType = "ACAL",
  backgroundPercentage = 40,
  qcPercentage = 80,
  nonReportableRSD = 30
)
}
\arguments{
\item{exp}{SummarizedExperiment Object}
}
\value{
SummarizedExperiment with added slots and assays with
batch-corrected values, including RSDQC, background signal, rt shift and
suggested internal standards.
}
\description{
This function is the main analysis function of mzQuality. It
starts by validating the SummarizedExperiment object is valid. Next, it will
calculate the batch-corrected RSDQC, background signals, retention time
shifts and lastly calculates a suggested internal standard that results in
the lowest batch-corrected RSDQC.
}
\details{
placeholder
}
\examples{
# Read the example dataset
exp <- readRDS(system.file("data.RDS", package = "mzQuality"))

# Perform analysis
exp <- doAnalysis(exp)
exp
}
