% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{downloadZip}
\alias{downloadZip}
\title{Download zip file}
\usage{
downloadZip(
  project,
  exp,
  summaryPlots = c("Aliquot", "PCA", "QC"),
  summaryReport = TRUE,
  compoundReport = TRUE,
  backgroundPercent = 40,
  cautionRSD = 15,
  nonReportableRSD = 30,
  copyDataLake = FALSE,
  assays = c("ratio", "ratio_corrected"),
  cores = 1,
  shinyProgress = NULL
)
}
\description{
Download zip file
}
