% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{exportExcel}
\alias{exportExcel}
\title{Export to Excel}
\usage{
exportExcel(
  exp,
  file,
  backgroundPercent = 40,
  cautionRSD = 15,
  nonReportableRSD = 30
)
}
\arguments{
\item{exp}{SummarizedExperiment Object}

\item{file}{Name of the file to export to}
}
\value{
placeholder

A workbook object from `openxlsx`
}
\description{
placeholder
}
\details{
placeholder
}
\examples{
# Read example dataset
data <- read.delim(system.file(package = "mzQuality", "dataset.txt"))

# Construct experiment
exp <- buildExperiment(
    data,
    rowIndex = "Compound",
    colIndex = "Aliquot",
    primaryAssay = "Area",
    secondaryAssay = "Area_is"
)

# Perform analysis
exp <- doAnalysis(exp)

# Request a summary in table format
exportExcel(exp, file = tempfile())
}
