% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{exportFilterWorkbook}
\alias{exportFilterWorkbook}
\title{Export Excel workbook}
\usage{
exportFilterWorkbook(exp, outfile = "Metabolites_to_keep.xlsx")
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{outfile}{File to export to}
}
\value{
placeholder

A workbook object from `openxlsx`
}
\description{
placeholder
}
\details{
placeholder
}
\examples{
# Read example dataset
data <- read.delim(system.file(package = "mzQuality", "dataset.txt"))

# Construct experiment
exp <- buildExperiment(
    data,
    rowIndex = "Compound",
    colIndex = "Aliquot",
    primaryAssay = "Area",
    secondaryAssay = "Area_is"
)

# Perform analysis
exp <- doAnalysis(exp)

# Export used aliquots and metabolites
exportFilterWorkbook(exp, outfile = tempfile())
}
