% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{exportTables}
\alias{exportTables}
\title{Export rowData, colData, and assays}
\usage{
exportTables(exp, folder)
}
\arguments{
\item{exp}{A SummarizedExperiment object}

\item{folder}{Folder where to store the tables}
}
\value{
placeholder

Creates files of rowData, colData and assays stored in `exp`
}
\description{
This function produces human-readable tables of rowData,
colData and assays of a SummarizedExperiment object. rowData will be stored
as 'Compounds.tsv', colData will be stored as 'Aliquots.tsv' and assays will
be stored under their assay name.
}
\details{
placeholder
}
\examples{
# Read example dataset
data <- read.delim(system.file(package = "mzQuality", "dataset.txt"))

# Construct experiment
exp <- buildExperiment(
    data,
    rowIndex = "Compound",
    colIndex = "Aliquot",
    primaryAssay = "Area",
    secondaryAssay = "Area_is"
)
# Do Analysis
exp <- doAnalysis(exp)

# Export tables in Experiment
exportTables(exp, folder = tempdir())
}
