% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundPlotNew.R
\name{facetPlot}
\alias{facetPlot}
\title{Facet a ggplot by a column}
\usage{
facetPlot(plot, by = "batch", ncol = 1, shareY = TRUE, shareX = FALSE)
}
\arguments{
\item{plot}{A ggplot2 object}

\item{by}{Character vector that matches one of the properties of the
`colData(exp)`. Defaults to `"batch"`}

\item{ncol}{Number of columns, defaults to 1}

\item{shareY}{Boolean value, should the y-axis be shared among facets?
Defaults to TRUE}

\item{shareX}{Boolean value, should the x-axis be shared among facets?
Defaults to FALSE}
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
