% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{fillMissingCombinations}
\alias{fillMissingCombinations}
\title{Fill in missing aliquot-compound combinations}
\usage{
fillMissingCombinations(df, rowIndex, colIndex, rowData, colData, assayNames)
}
\arguments{
\item{df}{Dataframe in long format}

\item{rowIndex}{Column name in `df` that should be used for compounds.}

\item{colIndex}{Column name in `df` that should be used for aliquots.}

\item{rowData}{Dataframe with rowData}

\item{colData}{Dataframe with colData}

\item{assayName}{Column name in `df` that represents the area / intensities
of measured compounds.}
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
