% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{filterBackground}
\alias{filterBackground}
\title{Filter the Background Signal}
\usage{
filterBackground(exp, min = 0, max = 0.4, type = "BLANK", include.na = FALSE)
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{min}{Minimum background signal value, defaults to 0}

\item{max}{Maximum background signal value, defaults to 0.4}

\item{type}{Aliquot type to use for background signal calculation. Defaults
to "BLANK"}

\item{include.na}{Should NA's be included?}
}
\value{
SummarizedExperiment with removed compounds based on the background
signal values and threshold provided
}
\description{
This function can be used to only keep compounds with a
low background signal.
}
\details{
placeholder
}
\examples{
# Read example dataset
data <- read.delim(system.file(package = "mzQuality", "dataset.txt"))

# Construct experiment
exp <- buildExperiment(
    data,
    rowIndex = "Compound",
    colIndex = "Aliquot",
    primaryAssay = "Area",
    secondaryAssay = "Area_is"
)

# Filter by background signal
filterBackground(exp, min = 0, max = 0.4)
}
