% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{filterISTD}
\alias{filterISTD}
\title{Filter compounds that are equal to given internal standards}
\usage{
filterISTD(exp, tag = "ISTD")
}
\arguments{
\item{exp}{A SummarizedExperiment object}

\item{tag}{Tag to search for in compound names that indicates an internal
standard. Defaults to 'ISTD'
Regular expressions are supported.}
}
\value{
SummarizedExperiment with removed compounds based on the filter
provided
}
\description{
In a scenario where internal standards were measured as
compounds, this function aids in removing these from the
SummarizedExperiment before doing analysis. It filter by both the `tag`
argument and by checking if the A/IS ratio equals 1.
}
\details{
placeholder
}
\examples{
# Read example dataset
data <- read.delim(system.file(package = "mzQuality", "dataset.txt"))

# Construct experiment
exp <- buildExperiment(
    data,
    rowIndex = "Compound",
    colIndex = "Aliquot",
    primaryAssay = "Area",
    secondaryAssay = "Area_is"
)

# Filter internal standards from compounds
filterISTD(exp, tag = "ISTD")
}
