% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{filterSST}
\alias{filterSST}
\title{Filter SST Aliquots from the SummarizedExperiment}
\usage{
filterSST(exp, tag = "SST")
}
\arguments{
\item{exp}{A SummarizedExperiment object}

\item{tag}{Tag to search for in aliquot types that indicate a SST.
Defaults to 'SST'
Regular expressions are supported.}
}
\value{
SummarizedExperiment with removed aliquots based on the filter
provided
}
\description{
System Sustainability Tests (SST) samples are often executed by
the technician for control during measurements. However, they are often
unwanted in analysis of measurements. This function helps to remove these
SSTs from the SummarizedExperiment object
}
\details{
This function is useful when you want to remove SST aliquots from
the analysis. It filters the aliquots based on the provided tag.
The tag is used to search for SST aliquots in the `type` column of the
SummarizedExperiment object. The function uses regular expressions to
match the tag, so you can use any pattern that matches the SST aliquots.
}
\examples{
path <- system.file("extdata/example.tsv", package = "mzQuality")
data <- readData(path)

# Construct experiment
exp <- buildExperiment(
    data,
    compoundColumn = "compound",
    aliquotColumn = "aliquot",
    primaryAssay = "area",
    secondaryAssay = "area_is"
)

# Filter SST aliquots
filterSST(exp, tag = "SST")
}
