% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{getConcentrationModelData}
\alias{getConcentrationModelData}
\title{Get Concentration Model Data}
\usage{
getConcentrationModelData(df, exp, calType, removeOutliers)
}
\arguments{
\item{df}{A data frame containing the data to be processed.}

\item{exp}{An experimental data object containing metadata and outlier
information.}

\item{calType}{A character string specifying the calibration type to
filter the data.}

\item{removeOutliers}{A logical value indicating whether to remove
outliers from the model data. Defaults to TRUE.}
}
\value{
A data frame containing the filtered and processed model data.
}
\description{
Prepares model data for concentration plots by filtering and
    optionally removing outliers.
}
\details{
This function filters the input data frame for a specific
    calibration type and optionally removes outliers based on the
    concentration outliers in the experimental data.
}
