% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{getConcentrationPlotData}
\alias{getConcentrationPlotData}
\title{Prepare Data for Concentration Plot}
\usage{
getConcentrationPlotData(
  exp,
  assay,
  batch,
  calType,
  types,
  plotOnCalibrationLine = TRUE
)
}
\arguments{
\item{exp}{A SummarizedExperiment object containing the experimental data.}

\item{assay}{A character string specifying the assay to be used for
plotting. Defaults to "concentration".}

\item{batch}{An integer specifying the batch index to filter the data.}

\item{calType}{A character string specifying the calibration type to include
in the plot. Defaults to the concentration metadata in `exp`.}

\item{types}{A character vector specifying the types of data to include
in the plot (e.g., "SAMPLE", "SQC").}

\item{plotOnCalibrationLine}{A logical value indicating whether to plot
data on the calibration line. Defaults to TRUE.}
}
\value{
A dataframe containing the processed data for the
    concentration plot.
}
\description{
Prepares the data required for generating a concentration plot,
    including calibration and sample data, with optional adjustments for
    plotting on the calibration line.
}
\details{
This function processes the experimental data to extract and
    structure the necessary information for creating a concentration plot.
    It supports filtering by batch and types, and optionally adjusts
    concentration values for specific types when not plotting on the
    calibration line. Additional annotations, such as text bubbles, are
    added to the data for enhanced visualization.
}
