% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{getMetaData}
\alias{getMetaData}
\title{Get columns that contain metadata in a long dataframe format}
\usage{
getMetaData(df, index)
}
\arguments{
\item{df}{A dataframe containing only columns with more than one unique
value.}

\item{index}{A string specifying the column to use as the index.}
}
\value{
A character vector of column names that contain metadata.
}
\description{
Identifies columns in a dataframe that contain metadata based
    on unique combinations of values in the specified index column.
}
\details{
This function checks each column in the dataframe (excluding the
    index column) to determine if it contains metadata. A column is
    considered metadata if it has unique values for each unique value in
    the index column.
}
