% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{getSuggestedInternalStandards}
\alias{getSuggestedInternalStandards}
\title{Suggest internal standards for compounds}
\usage{
getSuggestedInternalStandards(
  exp,
  secondaryAssay = metadata(exp)$secondary,
  removeOutliers = TRUE,
  useWithinBatch = TRUE
)
}
\arguments{
\item{exp}{A SummarizedExperiment object containing the experimental
data.}

\item{secondaryAssay}{A string specifying the assay to use for the
calculation. Defaults to the `secondary` metadata of the experiment.}

\item{removeOutliers}{Logical, whether to remove outliers during batch
correction. Defaults to `TRUE`.}

\item{useWithinBatch}{Logical, whether to use within-batch correction.
Defaults to `TRUE`.}
}
\value{
A SummarizedExperiment object with the suggested internal standards
    and their corresponding RSDQC values added to the `rowData`.
}
\description{
Identifies the internal standard that minimizes the
    batch-corrected RSDQC for each compound in a SummarizedExperiment
    object.
}
\details{
This function evaluates all available internal standards and
    suggests the one that results in the lowest batch-corrected RSDQC for
    each compound. The results are stored in the `rowData` of the
    experiment under the columns `suggestedIS` and `suggestedRSDQC`.
}
\examples{
# Example usage:
exp <- readRDS(system.file("extdata/data.RDS", package = "mzQuality"))
exp <- doAnalysis(exp, doAll = TRUE, removeOutliers = TRUE)
exp <- getSuggestedInternalStandards(exp, removeOutliers = TRUE)
}
