% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{heatmapPlot}
\alias{heatmapPlot}
\title{Plot heatmap of the given assay}
\usage{
heatmapPlot(exp, assay = "ratio")
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{assay}{Which assay to use for plotting the values}
}
\value{
Plotly heatmap of values in the provided assay
}
\description{
This functon plots a heatmap of the given assay. Constructed
using heatmaply, it is an interactive plot that shows the value of the
provide assay.
}
\examples{
# Read the example dataset
exp <- readRDS(system.file("data.RDS", package = "mzQuality2"))

# Plot values in heatmap
heatmapPlot(exp)
}
