% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{identifyMisInjections}
\alias{identifyMisInjections}
\title{Identify misinjections in Samples}
\usage{
identifyMisInjections(exp, assay = metadata(exp)$secondary, type = "SAMPLE")
}
\arguments{
\item{exp}{A SummarizedExperiment object containing the experimental data.}

\item{assay}{A string specifying the assay to use for outlier detection.
Defaults to the `secondary` metadata of the experiment.}

\item{type}{A string specifying the type of samples to analyze. Defaults
to `"SAMPLE"`.}
}
\value{
A SummarizedExperiment object with updated `colData` containing
    `use` and `outlier` columns.
}
\description{
Identifies misinjections in SAMPLEs for a specified assay in
    a SummarizedExperiment object.
}
\details{
This function is a wrapper around `identifyOutliers()` to detect
    misinjections in SAMPLEs based on the distribution of assay values.
    It updates the `colData` of the experiment with boolean columns `use`
    and `outlier`.
}
\examples{
# Example usage:
# exp <- readRDS(system.file("data.RDS", package = "mzQuality"))
# exp <- identifyMisInjections(exp, assay = "Ratio")
# Example usage:
exp <- readRDS(system.file("data.RDS", package = "mzQuality"))
exp <- identifyMisInjections(exp, assay = "area_is")
}
