% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{isValidExperiment}
\alias{isValidExperiment}
\title{Validate Experiment Object}
\usage{
isValidExperiment(exp)
}
\arguments{
\item{exp}{A SummarizedExperiment object.}
}
\value{
A boolean value indicating whether the SummarizedExperiment
  object is valid.
}
\description{
Validates if a SummarizedExperiment object is correctly
  formatted and suitable for use with the `mzQuality` package.
}
\details{
This function checks that the object is a valid
  SummarizedExperiment, contains the required metadata columns in
  `colData`, and includes the "ratio" assay.
}
\examples{
# Read example dataset
exp <- readRDS(system.file("extdata/data.RDS", package = "mzQuality"))

# Validate the experiment
isValid <- isValidExperiment(exp)
}
