% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSDQCs.R
\name{matrixRSDQCs}
\alias{matrixRSDQCs}
\title{Calculate the RSDQC of an assay}
\usage{
matrixRSDQCs(
  exp,
  primaryAssay = metadata(exp)$primary,
  secondaryAssay = metadata(exp)$secondary,
  qcType = metadata(exp)$QC
)
}
\arguments{
\item{exp}{SummarizedExperiment object}
}
\description{
The RSDQC is an indicator of variance of compounds in QC
samples. This function returns the RSDQC of all compounds in the
SummarizedExperiment
}
\details{
placeholder
}
\examples{
# Read the example dataset
exp <- readRDS(system.file("data.RDS", package = "mzQuality"))

matrixRSDQCs(exp)
}
