% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{medianSampleArea}
\alias{medianSampleArea}
\title{Calculate the median sample area for each compound}
\usage{
medianSampleArea(exp, type = "SAMPLE", assayName = metadata(exp)$primary)
}
\arguments{
\item{exp}{A SummarizedExperiment object containing the experimental data.}

\item{type}{A string specifying the type of samples to analyze. Defaults
to `"SAMPLE"`.}

\item{assayName}{A string specifying the assay to use for the calculation.
Defaults to the `primary` metadata of the experiment.}
}
\value{
A SummarizedExperiment object with the calculated median sample
    area added to the `rowData`.
}
\description{
Computes the median area for each compound in SAMPLEs for a
    specified assay in a SummarizedExperiment object.
}
\details{
This function calculates the median area for each compound in
    SAMPLEs based on the specified assay. The result is stored in the
    `rowData` of the experiment under the column `MedianSampleArea`.
}
