% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{multiApply}
\alias{multiApply}
\title{Apply jobs multicore}
\usage{
multiApply(
  func,
  jobs,
  cores = 1,
  progressTitle = "",
  shinyProgress = NULL,
  verbose = FALSE,
  ...
)
}
\description{
Apply jobs multicore
}
