% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQuality-package.R
\docType{package}
\name{mzQuality-package}
\alias{mzQuality}
\alias{mzQuality-package}
\title{mzQuality Package}
\description{
This is an R package for monitoring variation in mass
spectrometry measurements. It provides various metrics by utilizing pooled
quality control samples, blank samples and calibration lines. mzQuality uses
Bioconductors' SummarizedExperiment to facilitate implementation with
existing workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hankemeierlab/mzQuality}
  \item \url{https://hankemeierlab.github.io/mzQuality/Data_Input.html}
  \item Report bugs at \url{https://github.com/hankemeierlab/mzQuality/issues}
}

}
\author{
\strong{Maintainer}: Pascal Maas \email{p.maas@lacdr.leidenuniv.nl} (\href{https://orcid.org/0000-0001-9379-6041}{ORCID})

Authors:
\itemize{
  \item Alida Kindt \email{a.s.d.kindt@lacdr.leidenuniv.nl} (\href{https://orcid.org/0000-0001-6551-6030}{ORCID})
  \item Agnieszka Wegrzyn \email{a.b.wegrzyn@gmail.com} (\href{https://orcid.org/0000-0003-4872-8626}{ORCID})
}

Other contributors:
\itemize{
  \item Marielle van der Peet \email{m.p.van.der.peet@lacdr.leidenuniv.nl} (\href{https://orcid.org/0000-0001-8155-1485}{ORCID}) [contributor]
  \item Lieke Lamont \email{l.lamont@lacdr.leidenuniv.nl} (\href{https://orcid.org/0000-0002-2779-124X}{ORCID}) [contributor]
  \item Recon4IMD, Horizon Europe (101080997) [funder]
  \item SKINERGY, NWO (1389.20.182) [funder]
  \item X-Omics, NWO (184.034.019) [funder]
  \item EXPOSOME-SCAN, NWO (175.2019.032) [funder]
  \item EXPOSOME-NL, NWO (024.004.017) [funder]
}

}
\keyword{BatchEffect,}
\keyword{MassSpectrometry,}
\keyword{Metabolomics}
\keyword{QualityControl,}
\keyword{Software,}
\keyword{iternal,}
