% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQuality-package.R
\docType{package}
\name{mzQuality2-package}
\alias{mzQuality2}
\alias{mzQuality2-package}
\title{mzQuality Package}
\description{
This is an R package for monitoring technical variation in mass
spectrometry measurements. It provides various metrics by utilizing pooled
quality control samples, blank samples and calibration lines. mzQuality uses
Bioconductors' SummarizedExperiment to facilitate implementation with
existing workflows. It is fledged with an Rshiny application to support
interactive assessment of measurements.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{BatchEffect,}
\keyword{MassSpectrometry,}
\keyword{Metabolomics}
\keyword{QualityControl,}
\keyword{Software,}
\keyword{iternal,}
