% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{pcaPlot}
\alias{pcaPlot}
\title{PCA Plot of Metabolomics Data}
\usage{
pcaPlot(
  exp,
  assay = assayNames(exp)[1],
  batches = exp$batch,
  types = exp$type,
  pc1 = 1,
  pc2 = 2,
  logTransform = TRUE,
  sampleAsBatch = TRUE,
  addConfidenceInterval = TRUE,
  addLabels = FALSE
)
}
\arguments{
\item{exp}{A `SummarizedExperiment` object containing metabolomics data.}

\item{assay}{Character string specifying the assay to use for PCA
(default is the first assay in the object).}

\item{batches}{Character vector indicating which batches to include.
Defaults to all batches in the experiment.}

\item{types}{Character vector indicating which sample types to include.
Defaults to all types in the experiment.}

\item{pc1}{Integer indicating the principal component to plot on the
x-axis (default is 1).}

\item{pc2}{Integer indicating the principal component to plot on the
y-axis (default is 2).}

\item{logTransform}{Logical flag indicating whether to log-transform the
assay data before PCA (default is TRUE).}

\item{sampleAsBatch}{Logical flag. If TRUE, sample points are labeled
using batch numbers instead of type (default is TRUE).}

\item{addConfidenceInterval}{Logical flag. If TRUE, a 95% confidence
ellipse is drawn around each group (default is TRUE).}

\item{addLabels}{Logical, whether to add labels to the points in the plot.
requires `ggrepel` package. Defaults to `FALSE`.}
}
\value{
A `ggplot` object representing the PCA plot with options to
  customize axes, coloring, and annotation.
}
\description{
Performs principal component analysis (PCA) on
  metabolomics data and visualizes selected principal components
  with coloring and grouping by sample type or batch.
}
\details{
This function calculates PCA on selected assay data from a
  `SummarizedExperiment` object. Users can choose which principal
  components to plot, whether to log-transform the data, and whether
  to display confidence intervals. Sample types or batches can be
  visualized with custom colors. Optionally, sample names are shown
  using label bubbles for clarity.
}
\examples{
# Read data
exp <- readRDS(system.file("data.RDS", package = "mzQuality"))

# Do Analysis
exp <- doAnalysis(exp, doAll = TRUE)

# Create a compound plot for the first compound
pcaPlot(
   exp,
   assay = "ratio_corrected",
   logTransform = TRUE,
   types = c("SQC", "LQC", "SAMPLE")
)
}
