% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundPlotNew.R
\name{pcaPlotNew}
\alias{pcaPlotNew}
\title{New PCA Plot}
\usage{
pcaPlotNew(
  exp,
  assay = assayNames(exp)[1],
  batches = exp$batch,
  types = exp$type,
  pc1 = 1,
  pc2 = 2,
  doLog = TRUE,
  sampleAsBatch = TRUE,
  addConfidenceInterval = TRUE,
  typeColors = c(SQC = "black", LQC = "grey")
)
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
