% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{preparePlotData}
\alias{preparePlotData}
\title{Prepare plot data}
\usage{
preparePlotData(
  exp,
  assay = assayNames(exp)[1],
  batches = exp$batch,
  types = exp$type,
  doLog = FALSE,
  doImpute = FALSE
)
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
