% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{ratioQcSample}
\alias{ratioQcSample}
\title{Calculate the ratio between QCs and SAMPLEs}
\usage{
ratioQcSample(
  exp,
  assay = "ratio_corrected",
  qcLabel = metadata(exp)$QC,
  sampleLabel = "SAMPLE"
)
}
\arguments{
\item{exp}{A SummarizedExperiment object containing the experimental data.}

\item{assay}{A string specifying the assay to use for the calculation.
Defaults to `"ratio_corrected"`.}

\item{qcLabel}{A string specifying the label for QC samples. Defaults to
the `QC` metadata of the experiment.}

\item{sampleLabel}{A string specifying the label for SAMPLEs. Defaults to
`"SAMPLE"`.}
}
\value{
A SummarizedExperiment object with the calculated ratio added to
    the `rowData`.
}
\description{
Computes the ratio of median values between QC samples and
    SAMPLEs for a specified assay in a SummarizedExperiment object.
}
\details{
This function calculates the ratio of row medians for QC samples
    and SAMPLEs in the specified assay. The result is stored in the
    `rowData` of the SummarizedExperiment object under the column
    `ratioQcSample`.
}
\examples{
# Example usage:
# exp <- readRDS(system.file("data.RDS", package = "mzQuality"))
# exp <- ratioQcSample(exp)
}
