% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{readData}
\alias{readData}
\title{Construct a common Dataframe from vendor file(s)}
\usage{
readData(files, vendor = NA, regex = NULL)
}
\arguments{
\item{files}{Path to batch files or combined file}

\item{vendor}{optional Character, which vendor is used?}

\item{regex}{Regular Expression for parsing the aliquots}
}
\value{
placeholder

data.frame in a generic, common long format
}
\description{
Vendor files come in different export formats. This function
constructs a common, human-readable table format. It ensures that all
columnnames are titled, replaces any value equal to 'QC' to 'SQC' and adds
a batch number if none is present. Regex are used to parse aliquot names.
}
\details{
placeholder
}
