% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSDQCs.R
\name{replaceInternalStandards}
\alias{replaceInternalStandards}
\title{Replace internal standards in SummarizedExperiment}
\usage{
replaceInternalStandards(exp, internalStandards = rowData(exp)$suggestedIS)
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
\examples{
# Read the example dataset
exp <- readRDS(system.file("data.RDS", package = "mzQuality"))
}
