% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalLinesHighPerformance.R
\name{rowWiseCov}
\alias{rowWiseCov}
\title{Calculate covariance by row}
\usage{
rowWiseCov(x, y)
}
\arguments{
\item{x}{Matrix of X-values}

\item{y}{Matrix of y-values}
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
