% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalLinesHighPerformance.R
\name{rowWiseIntercept}
\alias{rowWiseIntercept}
\title{Calculate the intercept of a regression line by row}
\usage{
rowWiseIntercept(x, y, slope)
}
\arguments{
\item{x}{Matrix of X-values}

\item{y}{Matrix of y-values}

\item{slope}{Vector of named slope values, usually
obtained from [rowWiseSlope]}
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
