% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalLinesHighPerformance.R
\name{rowWiseSlope}
\alias{rowWiseSlope}
\title{Calculate the slope of a regression line by row}
\usage{
rowWiseSlope(x, y)
}
\arguments{
\item{x}{Matrix of X-values}

\item{y}{Matrix of y-values}
}
\description{
Calculate the slope of a regression line by row
}
