% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{setMetaDataFrame}
\alias{setMetaDataFrame}
\title{Set the Metadata of compound or aliquot data}
\usage{
setMetaDataFrame(rowData, rowIndex, rows)
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
