% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{setSampleColors}
\alias{setSampleColors}
\title{Set the colors of the samples}
\usage{
setSampleColors(types, ...)
}
\arguments{
\item{types}{Character vector with the types of the samples}

\item{...}{Additional arguments to be passed to the function}
}
\value{
placeholder
}
\description{
placeholder
details placeholder
}
