% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concentrations.R
\name{studentResiduals}
\alias{studentResiduals}
\title{Studentized residuals}
\usage{
studentResiduals(residuals, n, k = 1)
}
\arguments{
\item{residuals}{Matrix of residuals where each row is one compound and
the each column is an observation from the training set. This is usually
a CAL or ACAL line}

\item{n}{Number of observations. Should be equal to the amount of non-NA
values in the training set.}

\item{k}{Number of predictors. Here it is always 1}
}
\value{
Matrix of t-distributed residuals
}
\description{
Studentized residuals represent t-distributed residuals,
meaning that each residual is scored according to the t-distribution.
This allows for determining outliers in regression lines
}
\details{
This function effectively calculates the studentized residuals without
recalculating the model. The basic idea is to delete the observations one at
a time, each time refitting the regression model on the remaining n–1
observations. Then, we compare the observed response values to their
fitted values based on the models with the ith observation deleted. This
produces deleted residuals. Standardizing the deleted residuals produces
studentized residuals.
}
\seealso{
https://online.stat.psu.edu/stat462/node/247/ for the formula used here
}
