% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{summaryReport}
\alias{summaryReport}
\title{Create a dynamic mzQuality report}
\usage{
summaryReport(
  exp,
  folder,
  output = "mzquality-report.html",
  plots = c("Aliquot", "PCA", "QC"),
  assays = c("ratio", "ratio_corrected")
)
}
\arguments{
\item{exp}{SummarizedExperiment Object}

\item{folder}{Directory where to store the report.}

\item{output}{Output file}

\item{project}{Project name, defaults to "mzQuality"}

\item{vendor}{Vendor name of the files, defaults to "Unknown"}
}
\value{
placeholder

Creates a HTML file with dynamic visualizations
}
\description{
A summary report is a dynamic report containing plots involving
all compounds. Examples include a heatmap, aliquot plot and violin plot.
All plots are interactive and can be used for assessing experiments at a
later stage.
}
\details{
placeholder
}
\examples{
# Read example dataset
data(mzQualityExp)

# Request a summary report
# summaryReport(mzQualityExp,
#    folder = tempdir(), project = "MyProject",
#    vendor = "Agilent", output = tempfile()
#)
}
