% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{summaryTable}
\alias{summaryTable}
\title{Retrieve a summary of the SummarizedExperiment}
\usage{
summaryTable(exp, project = "mzQuality", vendor = "Unknown")
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{project}{Project name, defaults to "mzQuality"}

\item{vendor}{Vendor name of the files, defaults to "Unknown"}
}
\value{
placeholder

data.frame with a column "Info" and several rows of basic
information.
}
\description{
placeholder
}
\details{
placeholder
}
\examples{
# Read example dataset
data <- read.delim(system.file(package = "mzQuality", "dataset.txt"))

# Construct experiment
exp <- buildExperiment(
    data,
    rowIndex = "Compound",
    colIndex = "Aliquot",
    primaryAssay = "Area",
    secondaryAssay = "Area_is"
)

# Request a summary in table format
summaryTable(exp, project = "MyProject", vendor = "Sciex")
}
