% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{updateProgress}
\alias{updateProgress}
\title{Update the progressbar}
\usage{
updateProgress(pb, total, n, shinyProgress = NULL)
}
\arguments{
\item{pb}{Progressbar object}

\item{total}{Maximum value that the progressbar can have}

\item{n}{Current value of the progressbar}

\item{shinyProgress}{Optional identifier of a progressbar in shiny.}
}
\description{
This function aids in updating the progressbar for long-running
tasks.
}
\details{
When a shiny application is run and `shinyProgress` is given, this
function will also update the progress bar in the shiny app.
}
