% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{validateCalibration}
\alias{validateCalibration}
\title{Validate calibration dataframe}
\usage{
validateCalibration(exp, df)
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{df}{Dataframe with concentration measured in calibration lines}
}
\value{
Boolean value indicating if the calibration dataframe is suitable
for use with mzQuality
}
\description{
TO ADD
}
\examples{
# Read example dataset
data <- read.delim(system.file(package = "mzQuality", "dataset.txt"))

# Construct experiment
exp <- buildExperiment(
    data,
    rowIndex = "Compound",
    colIndex = "Aliquot",
    primaryAssay = "Area",
    secondaryAssay = "Area_is"
)
# Add calibration file + checks
}
