% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{validateDataframe}
\alias{validateDataframe}
\title{Validate measurements}
\usage{
validateDataframe(dataframe = NULL)
}
\arguments{
\item{dataframe}{data.frame with measurements.}
}
\value{
Boolean value if the dataframe is valid to use with mzQuality
}
\description{
placeholder
}
\details{
placeholder
}
\examples{
# Read example dataset
data <- read.delim(system.file(package = "mzQuality", "dataset.txt"))

# Validate if dataframe is corrct
validateDataframe(data)
}
