% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{validateExperiment}
\alias{validateExperiment}
\title{Validate experiment object}
\usage{
validateExperiment(exp)
}
\arguments{
\item{exp}{SummarizedExperiment object}
}
\value{
Boolean value indicating if the SummarizedExperiment is valid
to use with mzQuality
}
\description{
placeholder
}
\details{
placeholder
}
\examples{
# Read example dataset
data <- read.delim(system.file(package = "mzQuality", "dataset.txt"))

# Construct experiment
exp <- buildExperiment(
    data,
    rowIndex = "Compound",
    colIndex = "Aliquot",
    primaryAssay = "Area",
    secondaryAssay = "Area_is"
)

# Validate if the experiment is formatted correctly
validateExperiment(exp)
}
