% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{violinPlot}
\alias{violinPlot}
\title{Violin Plot}
\usage{
violinPlot(
  exp,
  assay = "ratio_corrected",
  batches = exp$batch,
  types = metadata(exp)$QC,
  doLog = FALSE,
  addMedian = TRUE,
  addConfidenceInterval = TRUE,
  withinTrend = FALSE
)
}
\arguments{
\item{exp}{An experimental data object containing the data to be plotted.}

\item{assay}{A character string specifying the assay column to be plotted
on the y-axis. Defaults to "ratio_corrected".}

\item{batches}{A vector specifying the batches to include in the plot.
Defaults to `exp$batch`.}

\item{types}{A vector specifying the types to include in the plot.
Defaults to `metadata(exp)$QC`.}

\item{doLog}{A logical value indicating whether to apply log scaling to
the y-axis. Defaults to FALSE.}

\item{addMedian}{A logical value indicating whether to add median points
to the plot. Defaults to TRUE.}

\item{addConfidenceInterval}{A logical value indicating whether to add
confidence intervals to the plot. Defaults to TRUE.}

\item{withinTrend}{A logical value indicating whether to inherit trend
line aesthetics. Defaults to FALSE.}
}
\value{
A ggplot object representing the violin plot.
}
\description{
Creates a violin plot with optional median points and
    confidence intervals.
}
\details{
This function generates a violin plot using ggplot2. It allows
    for optional log scaling, adding median points, and confidence
    intervals.
}
