% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundPlotNew.R
\name{violinPlotNew}
\alias{violinPlotNew}
\title{New Violin Plot}
\usage{
violinPlotNew(
  exp,
  assay = "ratio_corrected",
  batches = exp$batch,
  types = metadata(exp)$QC,
  doLog = FALSE,
  addMedian = TRUE,
  addConfidenceInterval = TRUE,
  withinTrend = FALSE
)
}
\value{
placeholder
}
\description{
A short description...
}
\details{
placeholder
}
