% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{zipFolder}
\alias{zipFolder}
\title{Zip the folder with results}
\usage{
zipFolder(file, output)
}
\arguments{
\item{file}{output name of file}

\item{output}{Folder to be zipped}
}
\value{
placeholder

A zip file is created of the folders created with the export
functions and reports.
}
\description{
placeholder
}
\details{
placeholder
}
