% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{addConcentrations}
\alias{addConcentrations}
\title{Add known concentrations of calibration lines}
\usage{
addConcentrations(exp, df, filterComps = FALSE)
}
\arguments{
\item{exp}{SummarizedExperiment to add the data to}

\item{df}{Dataframe with concentrations of CAL lines}

\item{filterComps}{Boolean value, should compounds without concentration
be removed?}
}
\value{
SummarizedExperiment with an added assay called "Concentrations".
The values of the assay are filled with the values provided in `df`, other
values will be `NA`.
}
\description{
mzQuality supports the calculation of absolute concentrations
given concentrations of Calibration lines (CAL) are present. Using this
function, a new assay in the experiment is created called 'Concentration',
which is filled with NA and the supplied concentrations.

The concentrations should be supplied as a data frame with rows as the
names of Calibration lines and columns as compound names.
Optionally, the `filterComps` argument can be set to TRUE, which only keeps
compounds in the experiment with a known concentration in the CAL lines.
}
