% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concentrations.R
\name{addLinearRange}
\alias{addLinearRange}
\title{Add the ratio of compounds present in the linear range}
\usage{
addLinearRange(
  exp,
  assay = "ratio",
  calType = metadata(exp)$concentration,
  minCalNo = 2,
  maxCalNo = 6,
  saveAssay = "CalRange"
)
}
\arguments{
\item{exp}{A SummarizedExperiment object containing the experimental data.}

\item{assay}{A string specifying the assay name to use for ratio values.
Defaults to "ratio".}

\item{calType}{A string specifying the calibration type to use. Defaults to
the concentration metadata of the experiment.}

\item{minCalNo}{An integer specifying the minimum calibration number to
include in the range. Defaults to 2.}

\item{maxCalNo}{An integer specifying the maximum calibration number to
include in the range. Defaults to 6.}

\item{saveAssay}{A string specifying the name of the assay to save the
boolean matrix. Defaults to "CalRange".}
}
\value{
A SummarizedExperiment object with the updated assay containing the
    boolean matrix of compounds within the linear range.
}
\description{
Adds a boolean matrix to the experiment object indicating
    whether the ratios of compounds fall within the specified calibration
    range for each batch.
}
\details{
This function checks whether the ratios of compounds in the
    experiment object fall within the specified calibration range (defined
    by `minCalNo` and `maxCalNo`) for each batch. It creates a boolean
    matrix indicating whether each compound is within the range and saves
    it as a new assay in the experiment object.
}
\examples{
# Example usage:
exp <- readRDS(system.file("data.RDS", package = "mzQuality"))
exp <- addLinearRange(
    exp, assay = "ratio", minCalNo = 2,
    maxCalNo = 6, saveAssay = "CalRange"
)
}
