% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchEffects.R
\name{addWithinBatchCorrectionAssay}
\alias{addWithinBatchCorrectionAssay}
\title{Within-Batch correction}
\usage{
addWithinBatchCorrectionAssay(
  exp,
  assay = "ratio",
  saveAssay = "ratio_corrected",
  qcType = metadata(exp)$QC,
  removeOutliers = TRUE
)
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
