% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{aliquotPlot}
\alias{aliquotPlot}
\title{Aliquot Plot}
\usage{
aliquotPlot(exp, assay, batches = exp$batch, types = exp$type, doLog = FALSE)
}
\value{
placeholder
}
\description{
A short description...
}
\details{
placeholder
}
