% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalLinesHighPerformance.R
\name{batchConcentrations}
\alias{batchConcentrations}
\title{Calculate concentrations per batch}
\usage{
batchConcentrations(
  exp,
  batch,
  assayX = "ratio",
  assayY = "concentration",
  type = "ACAL",
  subtractCal = FALSE,
  forceOrigin = FALSE,
  checkForOutliers = TRUE,
  subtractIntercept = FALSE,
  useWeights = FALSE
)
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{checkForOutliers}{Boolean value, should studentized residuals be
used to determine outliers? If TRUE, outliers are detected and only removed
at the start or end of a calibration line. If FALSE, outlier detection
is ignored.}

\item{useWeights}{Boolean value, should weighted regression be used? Defaults
to FALSE, which indicates no weights are used.}
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
