% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{buildExperiment}
\alias{buildExperiment}
\title{Create a SummarizedExperiment object from a long dataframe}
\usage{
buildExperiment(
  df,
  rowIndex = "compound",
  colIndex = "aliquot",
  primaryAssay = "area",
  secondaryAssay = "area_is",
  typeColumn = "type",
  qc = "SQC",
  secondaryIndex = "compound_is"
)
}
\arguments{
\item{df}{Dataframe in a long format. Should contain at least a column for
compounds, aliquots, area, type and datetime. See the vignette for an
example of a valid dataframe.}

\item{rowIndex}{Column name in `df` that should be used for compounds.
Defaults to 'Compound'}

\item{colIndex}{Column name in `df` that should be used for aliquots.
Defaults to 'Aliquot'}

\item{primaryAssay}{Column name in `df` that represents the
area / intensities of measured compounds. Defaults to 'Area'}

\item{secondaryAssay}{Column name in `df` that represents the
area / intensities of measured internal standards.
If no internal standards are present, this parameter can be omitted and the
 A/IS ratio will be set to 1.}

\item{qc}{Name of the QC sample used for batch correction found in the
'Type' column of `df`. Defaults to 'SQC'}
}
\value{
A SummarizedExperiment object with dimensions compounds x aliquots.
}
\description{
mzQuality works with a SummarizedExperiment object to perform
its analysis. This function aids in that process by constructing one for
you. It uses cardinality to identify in which slot a column should be
stored. After creation, it will validate if the object is suitable for
mzQuality to use.
}
\examples{
# Read example dataset
data <- readData(system.file(package = "mzQuality", "example.tsv"))

# Construct experiment
exp
}
