% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalLinesHighPerformance.R
\name{calculateConcentrations}
\alias{calculateConcentrations}
\title{Calculate concentratios for a full experiment}
\usage{
calculateConcentrations(exp, byBatch = TRUE, saveAssay = "concentration", ...)
}
\arguments{
\item{byBatch}{Should the concentration be calculated per batch? Defaults to
TRUE.}

\item{saveAssay}{Name of the assay to save the results. Defaults to
`"Concentration"`}
}
\value{
SummarizedExperiment with predicted concentrations in the assay
given in `saveAssay`. Defaults to `"concentration"`.
}
\description{
Calculate concentratios for a full experiment
}
