% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{compoundPlot}
\alias{compoundPlot}
\title{Compound Plot}
\usage{
compoundPlot(
  exp,
  compound = 1,
  batches = exp$batch,
  types = exp$type,
  assay = "ratio_corrected",
  logTransform = TRUE,
  withinTrend = FALSE,
  shortName = FALSE,
  trendTypes = metadata(exp)$QC,
  addInternalStandards = FALSE,
  addText = FALSE
)
}
\arguments{
\item{exp}{A `SummarizedExperiment` object containing
metabolomics data.}

\item{compound}{Integer or character. Row index or name of
the compound to be plotted.}

\item{batches}{Character vector of batch identifiers to include
(default is all batches in `exp`).}

\item{types}{Character vector of sample types to include
(default is all types in `exp`).}

\item{assay}{Character string specifying the assay to use
(default is "ratio_corrected").}

\item{logTransform}{Logical. Whether to log-transform the y-axis
values (default is TRUE).}

\item{withinTrend}{Logical. If TRUE, trendlines are drawn per
batch within types (default is FALSE).}

\item{shortName}{Logical. If TRUE, short sample names are used
on the x-axis (default is FALSE).}

\item{trendTypes}{Character vector of sample types to be used
for trendline fitting (default is `metadata(exp)$QC`).}

\item{addInternalStandards}{Logical. Whether to add
corresponding internal standards (if available)
(default is FALSE).}

\item{addText}{Logical. Whether to add text bubbles to the
plot, only visible when using plotly (default is FALSE).}
}
\value{
A `ggplot` object showing compound signal
  intensities over time, grouped and colored by type or batch.
}
\description{
Creates a line plot of intensity or ratio values for a
  specific compound across injections, colored by sample
  type or batch, with optional trendlines and internal
  standard visualization.
}
\details{
This function generates a time-series-like plot for
  a specific compound in a `SummarizedExperiment` object. It
  includes support for multiple batches, sample types,
  internal standards, trend smoothing, and optional use of
  short sample names or log transformation.
}
\examples{
# Read data
exp <- readRDS(system.file("data.RDS", package = "mzQuality"))

# Do Analysis
exp <- doAnalysis(exp, doAll = TRUE)

# Create a compound plot for the first compound
compoundPlot(
   exp,
   compound = 1,
   types = c("SQC", "LQC"),
   assay = "ratio_corrected",
   logTransform = TRUE
)
}
