% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{compoundReports}
\alias{compoundReports}
\title{Create dynamic compound report(s)}
\usage{
compoundReports(
  exp,
  folder,
  assays = c("ratio", "ratio_corrected"),
  project = "mzQuality",
  cores = 1,
  verbose = F,
  shinyProgress = NULL
)
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{folder}{Folder where to store the compound report.}

\item{project}{Project name, defaults to "mzQuality"}

\item{compounds}{Which compounds to create a report for. Defaults to all
compounds in `exp`}

\item{vendor}{Vendor name of the files, defaults to "Unknown"}

\item{output}{Output file}
}
\value{
placeholder

Creates a HTML file with dynamic visualizations
}
\description{
A compound report is a dynamic report containing plots
regarding a single compound. By default, this function produces a plot for
each compound, but can also be made for a single compound.
}
\details{
placeholder
}
\examples{
# Read example dataset
data <- read.delim(system.file(package = "mzQuality", "dataset.txt"))

# Construct experiment
exp <- buildExperiment(
    data,
    rowIndex = "Compound",
    colIndex = "Aliquot",
    primaryAssay = "Area",
    secondaryAssay = "Area_is"
)

# Perform analysis
exp <- doAnalysis(exp)

# Request a compound report
}
