% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{convertExperiment}
\alias{convertExperiment}
\title{Convert an experiment to one capable with mzQuality}
\usage{
convertExperiment(
  exp,
  primaryAssay = "area",
  secondaryAssay = primaryAssay,
  istd = "compound_is",
  type = "type",
  datetime = "datetime",
  batch = "batch",
  qcType = "SQC"
)
}
\arguments{
\item{exp}{A SummarizedExperiment object}

\item{primaryAssay}{Assay that should be used for compound Areas / Intensity}

\item{secondaryAssay}{Assay that should be used for internal standard
Areas / Intensity. Defaults to equal to primaryAssay, meaning no internal
standards are present.}

\item{istd}{Column in the rowData that represents the internal standard used
for the compound. Defaults to 'compound_is'}

\item{type}{Column in the colData that represents the sample type.
Defaults to 'Type'}

\item{datetime}{Column in the colData that represents the datetime when the
sample was measured. Defaults to 'Datetime'}

\item{batch}{Column in the colData that represents the batch name / number.
Defaults to 'Batch'}

\item{qcType}{Name of the QC sample type used. Defaults to 'SQC'}
}
\value{
SummarizedExperiment that is suitable for use with mzQuality
}
\description{
This function can be used to convert an existing
SummarizedExperiment to one capable to use with mzQuality. This function
will change some colData and rowData columns, but will leave assays
untouched.
}
\examples{
# Read example dataset
data <- read.delim(system.file(package = "mzQuality", "dataset.txt"))

# Construct experiment
exp <- buildExperiment(
    data,
    rowIndex = "Compound",
    colIndex = "Aliquot",
    primaryAssay = "Area",
    secondaryAssay = "Area_is"
)
convertExperiment(exp, primaryAssay = "Area")
}
