% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{facetPlot}
\alias{facetPlot}
\title{Facet a ggplot by a column}
\usage{
facetPlot(plot, by = "batch", ncol = 1, shareY = TRUE, shareX = FALSE)
}
\arguments{
\item{plot}{A ggplot2 object to be faceted.}

\item{by}{A character vector specifying the column to facet by. Defaults
to "batch".}

\item{ncol}{An integer specifying the number of columns in the facet.
Defaults to 1.}

\item{shareY}{A logical value indicating whether the y-axis should be
shared among facets. Defaults to TRUE.}

\item{shareX}{A logical value indicating whether the x-axis should be
shared among facets. Defaults to FALSE.}
}
\value{
A faceted ggplot object.
}
\description{
Facets a ggplot object by a specified column.
}
\details{
This function uses ggplot2's facet_wrap to create facets for a
    ggplot object based on a specified column. It allows customization of
    the number of columns and whether axes are shared.
}
\examples{
# Read data
exp <- readRDS(system.file("data.RDS", package = "mzQuality"))

# Do Analysis
exp <- doAnalysis(exp, doAll = TRUE)

# Create a compound plot for the first compound
p <- compoundPlot(
   exp,
   compound = 1,
   types = c("SQC", "LQC"),
   assay = "ratio_corrected",
   logTransform = TRUE
)

facetPlot(p, by = "batch", ncol = 2, shareY = TRUE)
}
