% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{fillMissingCombinations}
\alias{fillMissingCombinations}
\title{Fill Missing Aliquot-Compound Combinations}
\usage{
fillMissingCombinations(df, rowIndex, colIndex, rowData, colData, assayNames)
}
\arguments{
\item{df}{A dataframe in long format.}

\item{rowIndex}{The column name in `df` to be used for compounds.}

\item{colIndex}{The column name in `df` to be used for aliquots.}

\item{rowData}{A dataframe containing row data (e.g., compounds).}

\item{colData}{A dataframe containing column data (e.g., aliquots).}

\item{assayNames}{A character vector of column names in `df` representing
the area or intensities of measured compounds.}
}
\value{
A dataframe with all aliquot-compound combinations, including
  the missing ones filled with NA values.
}
\description{
Identifies and fills in missing combinations of aliquots
  and compounds in a long-format dataframe.
}
\details{
This function ensures that all possible combinations of aliquots
  and compounds are present in the dataframe. Missing combinations are
  added with NA values for the assay columns.
}
