% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{filterBackground}
\alias{filterBackground}
\title{Filter the Background Signal}
\usage{
filterBackground(exp, min = 0, max = 0.4, include.na = FALSE)
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{min}{Minimum background signal value, defaults to 0}

\item{max}{Maximum background signal value, defaults to 0.4}

\item{include.na}{Should NA's be included?}
}
\value{
SummarizedExperiment with removed compounds based on the background
signal values and threshold provided
}
\description{
This function can be used to only keep compounds with a
low background signal.
}
\details{
This function is useful when you want to filter compounds based on
the background signal values. It filters the compounds based on the
provided minimum and maximum background signal values.
}
\examples{
path <- system.file("extdata", "example.tsv", package = "mzQuality")
data <- readData(path)

# Construct experiment
exp <- buildExperiment(
    data,
    compoundColumn = "compound",
    aliquotColumn = "aliquot",
    primaryAssay = "area",
    secondaryAssay = "area_is"
)

# Filter by background signal
filterBackground(exp, min = 0, max = 0.4)
}
