% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{filterISTD}
\alias{filterISTD}
\title{Filter compounds that are equal to given internal standards}
\usage{
filterISTD(exp, tag = "STD")
}
\arguments{
\item{exp}{A SummarizedExperiment object}

\item{tag}{Tag to search for in compound names that indicates an internal
standard. Defaults to 'ISTD'
Regular expressions are supported.}
}
\value{
SummarizedExperiment with removed compounds based on the filter
provided
}
\description{
In a scenario where internal standards were measured as
compounds, this function aids in removing these from the
SummarizedExperiment before doing analysis. It filter by both the `tag`
argument and by checking if the A/IS ratio equals 1.
}
\details{
This function is useful when internal standards are
measured as compounds and you want to remove them from the analysis.
It filters the compounds based on the provided tag and checks if the A/IS
ratio equals 1. If the `hasIS` metadata is set to TRUE, it will also
remove compounds that have the same A/IS ratio as the internal standard.
}
\examples{
path <- system.file("extdata/example.tsv", package = "mzQuality")
data <- readData(path)

# Construct experiment
exp <- buildExperiment(
    data,
    compoundColumn = "compound",
    aliquotColumn = "aliquot",
    primaryAssay = "area",
    secondaryAssay = "area_is"
)

# Filter internal standards from compounds
filterISTD(exp, tag = "ISTD")

# Filter all standards from compounds
filterISTD(exp, tag = "STD")
}
