% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{filterOutliers}
\alias{filterOutliers}
\title{Filter aliquot QC outliers from the dataset}
\usage{
filterOutliers(exp)
}
\arguments{
\item{exp}{SummarizedExperiment object}
}
\value{
SummarizedExperiment with removed QC aliquots based their median
values
}
\description{
This function identifies and removes outlier QC aliquots from
the Experiment. Outliers are found by applying the RosnerTest on the median
A/IS ratio of QCs.

placeholder
}
\examples{
# Read example dataset
exp <- readRDS(system.file("extdata/data.RDS", package = "mzQuality"))

# Remove QC outliers
filterOutliers(exp)
}
