% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{filterOutliers}
\alias{filterOutliers}
\title{Filter aliquot QC outliers from the dataset}
\usage{
filterOutliers(exp)
}
\arguments{
\item{exp}{SummarizedExperiment object}
}
\value{
SummarizedExperiment with removed QC aliquots based their median
values
}
\description{
This function identifies and removes outlier QC aliquots from
the Experiment. Outliers are found by applying the RosnerTest on the median
A/IS ratio of QCs.

placeholder
}
\examples{
# Read example dataset
data <- readData(system.file(package = "mzQuality", "example.tsv"))

# Construct experiment
exp <- buildExperiment(
    data,
    rowIndex = "compound",
    colIndex = "aliquot",
    primaryAssay = "area",
    secondaryAssay = "area_is"
)

# Remove QC outliers
filterOutliers(exp)
}
