% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{filterRSDQC}
\alias{filterRSDQC}
\title{Filter the compounds based on batch-corrected RSDQC}
\usage{
filterRSDQC(exp, min = 0, max = 30, include.na = FALSE)
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{min}{Minimum RSDQC value, defaults to 0}

\item{max}{Maximum RSDQC value, defaults to 30}

\item{include.na}{Should NA's be included?}
}
\value{
SummarizedExperiment with removed compounds based on the RSDQC
threshold provided
}
\description{
This function can be used to only keep compounds with
a RSDQC threshold. This ensures that only reliably measured compounds
are kept.
}
\details{
This function is useful when you want to filter compounds based on
the RSDQC values. It filters the compounds based on the provided
minimum and maximum RSDQC values. The function uses regular expressions
to match the tag, so you can use any pattern that matches the RSDQC
values.
}
\examples{
# Read example dataset
exp <- readRDS(system.file(package = "mzQuality", "data.RDS"))

# Perform analysis
exp <- doAnalysis(exp)

# Keep compounds with a maximum RSDQC of 30
filterRSDQC(exp, min = 0, max = 30)

# Keep compounds between RSDQC values 5 and 15
filterRSDQC(exp, min = 5, max = 15)
}
