% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{filterSST}
\alias{filterSST}
\title{Filter SST Aliquots from the SummarizedExperiment}
\usage{
filterSST(exp, tag = "SST")
}
\arguments{
\item{exp}{A SummarizedExperiment object}

\item{tag}{Tag to search for in aliquot types that indicate a SST.
Defaults to 'SST'
Regular expressions are supported.}
}
\value{
SummarizedExperiment with removed aliquots based on the filter
provided
}
\description{
System Sustainability Tests (SST) samples are often executed by
the technician for control during measurements. However, they are often
unwanted in analysis of measurements. This function helps to remove these
SSTs from the SummarizedExperiment object
}
\details{
placeholder
}
\examples{
# Read example dataset
data <- read.delim(system.file(package = "mzQuality", "dataset.txt"))

# Construct experiment
exp <- buildExperiment(
    data,
    rowIndex = "Compound",
    colIndex = "Aliquot",
    primaryAssay = "Area",
    secondaryAssay = "Area_is"
)

# Filter SST aliquots
filterSST(exp, tag = "SST")
}
