% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{getMetaData}
\alias{getMetaData}
\title{Get columns that contain metadata in a long dataframe format}
\usage{
getMetaData(df, index)
}
\arguments{
\item{df}{Dataframe with only columns with > 1 unique values}

\item{index}{Which index should be used as the index column}
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
