% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundPlotNew.R
\name{getPcaData}
\alias{getPcaData}
\title{Transform data into PCA data}
\usage{
getPcaData(
  exp,
  assay = assayNames(exp)[1],
  batches = exp$batch,
  types = exp$type,
  doLog = TRUE
)
}
\value{
placeholder
}
\description{
A short description...
}
\details{
placeholder
}
