% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concentrations.R
\name{getWeights}
\alias{getWeights}
\title{Get weights for weighted regression}
\usage{
getWeights(yValues)
}
\arguments{
\item{yValues}{Values of the response}
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
