% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{identifyMisInjections}
\alias{identifyMisInjections}
\title{Identifier misinjections in Samples}
\usage{
identifyMisInjections(exp, assay = metadata(exp)$secondary, type = "SAMPLE")
}
\arguments{
\item{exp}{SummarizedExperiment Object}

\item{assay}{Which assay should be used for outlier detection}
}
\value{
placeholder

SummarizedExperiment with added column called "Use" with boolean
values suggesting if it should be used in analysis
}
\description{
placeholder
}
\details{
placeholder
}
\examples{
# Read the example dataset
exp <- readRDS(system.file("data.RDS", package = "mzQuality2"))

# Identify aliquot outliers and add to colData
exp <- identifyOutliers(exp, assay = "Ratio")
}
