% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{identifyOutliers}
\alias{identifyOutliers}
\title{Identify outliers in QC samples}
\usage{
identifyOutliers(exp, assay = "ratio", qcType = metadata(exp)$QC)
}
\arguments{
\item{exp}{A SummarizedExperiment object containing the experimental data.}

\item{assay}{A string specifying the assay to use for outlier detection.
Defaults to `"ratio"`.}

\item{qcType}{A string specifying the type of QC samples. Defaults to the
`QC` metadata of the experiment.}
}
\value{
A SummarizedExperiment object with updated `colData` containing
    `use` and `outlier` columns.
}
\description{
Identifies outliers in QC samples for a specified assay in a
    SummarizedExperiment object and updates the `colData` with outlier
    information.
}
\details{
This function detects outliers in QC samples based on the
    distribution of log-transformed assay values. It uses Rosner's test to
    identify outliers and updates the `colData` of the experiment with
    boolean columns `use` and `outlier`. Outliers are marked as `FALSE` in
    the `use` column and `TRUE` in the `outlier` column.
}
\examples{
# Example usage:
# exp <- readRDS(system.file("data.RDS", package = "mzQuality"))
# exp <- identifyOutliers(exp, assay = "ratio")
}
