% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlankEffects.R
\name{matrixEffect}
\alias{matrixEffect}
\title{Calculate the matrix effect and ion suppression}
\usage{
matrixEffect(
  exp,
  is_assay = metadata(exp)$secondary,
  sampleLabel = "SAMPLE",
  procBlankLabel = "PROC"
)
}
\arguments{
\item{exp}{SummarizedExperiment}
}
\description{
Calculate the matrix effect and ion suppression
}
\examples{
#' # Read the example dataset
exp <- readRDS(system.file("data.RDS", package = "mzQuality"))

matrixEffect(exp)
}
