% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSDQCs.R
\name{matrixRSDQCs}
\alias{matrixRSDQCs}
\title{Calculate the RSDQC of an assay}
\usage{
matrixRSDQCs(
  exp,
  primaryAssay = metadata(exp)$primary,
  secondaryAssay = metadata(exp)$secondary,
  qcType = metadata(exp)$QC
)
}
\arguments{
\item{exp}{A SummarizedExperiment object containing the experimental data.}

\item{primaryAssay}{A string specifying the name of the primary assay to
use. Defaults to the `primary` metadata of the experiment.}

\item{secondaryAssay}{A string specifying the name of the secondary assay
to use. Defaults to the `secondary` metadata of the experiment.}

\item{qcType}{A string specifying the type of QC samples. Defaults to the
`QC` metadata of the experiment.}
}
\value{
A matrix of RSDQC values, with rows corresponding to internal
    standards and columns corresponding to compounds.
}
\description{
Computes the Relative Standard Deviation for Quality Control
    (RSDQC) of compounds in QC samples from a SummarizedExperiment object.
}
\details{
This function calculates the RSDQC, which is an indicator of the
    variance of compounds in QC samples. It supports both primary and
    secondary assays and performs between-batch correction if internal
    standards (IS) are available. The result is a matrix of RSDQC values
    for each compound and internal standard.
}
\examples{
# Example usage:
# exp <- readRDS(system.file("extdata/data.RDS", package = "mzQuality"))
# rsdMatrix <- matrixRSDQCs(exp)
# Example usage:
exp <- readRDS(system.file("extdata/data.RDS", package = "mzQuality"))

# Calculate the RSDQC matrix for the default primary and secondary assays
rsd_matrix <- matrixRSDQCs(exp)
}
