% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQuality-package.R
\docType{package}
\name{mzQuality-package}
\alias{mzQuality}
\alias{mzQuality-package}
\title{mzQuality Package}
\description{
This is an R package for monitoring technical variation in mass
spectrometry measurements. It provides various metrics by utilizing pooled
quality control samples, blank samples and calibration lines. mzQuality uses
Bioconductors' SummarizedExperiment to facilitate implementation with
existing workflows. It is fledged with an Rshiny application to support
interactive assessment of measurements.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hankemeierlab/mzQuality}
  \item Report bugs at \url{https://github.com/hankemeierlab/mzQuality/issues}
}

}
\author{
\strong{Maintainer}: Pascal Maas \email{p.maas@lacdr.leidenuniv.nl} (\href{https://orcid.org/0000-0001-9379-6041}{ORCID})

Authors:
\itemize{
  \item Alida Kindt \email{a.s.d.kindt@lacdr.leidenuniv.nl} (\href{https://orcid.org/0000-0001-6551-6030}{ORCID})
  \item Agnieszka Wegrzyn (\href{https://orcid.org/0000-0003-4872-8626}{ORCID})
}

Other contributors:
\itemize{
  \item Marielle van der Peet \email{m.p.van.der.peet@lacdr.leidenuniv.nl} (\href{https://orcid.org/0000-0001-8155-1485}{ORCID}) [contributor]
  \item Lieke Lamont \email{l.lamont@lacdr.leidenuniv.nl} (\href{https://orcid.org/0000-0002-2779-124X}{ORCID}) [contributor]
  \item Recon4IMD [funder]
  \item NGID [funder]
  \item EXPOSOME-NL [funder]
}

}
\keyword{BatchEffect,}
\keyword{MassSpectrometry,}
\keyword{Metabolomics}
\keyword{QualityControl,}
\keyword{Software,}
\keyword{iternal,}
