% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{ratioQcSample}
\alias{ratioQcSample}
\title{Calculate the ratio between QCs and SAMPLEs}
\usage{
ratioQcSample(
  exp,
  assay = "ratio_corrected",
  qcLabel = metadata(exp)$QC,
  sampleLabel = "SAMPLE"
)
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
\examples{
# Read the example dataset
exp <- readRDS(system.file("data.RDS", package = "mzQuality"))

# Calculate the ratio between QCs and Samples of the batch-corrected ratio
ratioQcSample(exp)
}
