% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{readData}
\alias{readData}
\title{Construct a Common Dataframe from Vendor File(s)}
\usage{
readData(files, vendor = NA, regex = NULL)
}
\arguments{
\item{files}{A character vector of paths to batch files or a combined file.}

\item{vendor}{An optional character string specifying the vendor used.}

\item{regex}{A regular expression for parsing aliquot names.}
}
\value{
A data.frame in a generic, common long format with standardized
  column names and additional metadata such as batch numbers.
}
\description{
This function processes vendor files, which may come in
  different export formats, and constructs a common, human-readable
  table format. It ensures that all column names are standardized,
  replaces any value equal to 'QC' with 'SQC', and adds a batch number
  if none is present. Regular expressions are used to parse aliquot names.
}
\details{
The function handles vendor-specific formats by applying
  appropriate parsing routines. For Sciex files, it uses a specific
  routine to process the data. Missing values are replaced with `NA`,
  and column names are converted to lowercase for consistency.
}
