% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSDQCs.R
\name{rsdqc}
\alias{rsdqc}
\title{Calculate the RSDQC of an assay}
\usage{
rsdqc(exp, assay = "ratio_corrected", type = metadata(exp)$QC)
}
\arguments{
\item{exp}{SummarizedExperiment object}

\item{assay}{Assay to use for calculating the RSDQC. Defaults to
"ratio_corrected"}
}
\description{
The RSDQC is an indicator of variance of compounds in QC
samples. This function returns the RSDQC of all compounds in the
SummarizedExperiment
}
\details{
placeholder
}
\examples{
# Read the example dataset
exp <- readRDS(system.file("data.RDS", package = "mzQuality2"))

rsdqc(exp)
}
