% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Compound plot function}
\usage{
scatterPlot(df, assay, trendTypes, colors, smooth = TRUE, doLog = TRUE)
}
\value{
placeholder
}
\description{
A short description...
}
\details{
placeholder
}
