% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{validateDataframe}
\alias{validateDataframe}
\title{Validate Measurements}
\usage{
validateDataframe(dataframe = NULL)
}
\arguments{
\item{dataframe}{A data.frame containing measurements.}
}
\value{
A boolean value indicating whether the dataframe is valid.
}
\description{
Validates if a dataframe is suitable for use with the
  `mzQuality` package.
}
\details{
This function checks if the dataframe contains the required
  columns, has no duplicate aliquot-batch-compound combinations, and
  includes at least one "QC" type.
}
\examples{
# Read example dataset
data <- readData(system.file(package = "mzQuality", "example.tsv"))

# Validate if the dataframe is correct
validateDataframe(data)
}
