% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{validateExperiment}
\alias{validateExperiment}
\title{Validate Experiment Object}
\usage{
validateExperiment(exp)
}
\arguments{
\item{exp}{A SummarizedExperiment object.}
}
\value{
A boolean value indicating whether the SummarizedExperiment
  object is valid.
}
\description{
Validates if a SummarizedExperiment object is correctly
  formatted and suitable for use with the `mzQuality` package.
}
\details{
This function checks that the object is a valid
  SummarizedExperiment, contains the required metadata columns in
  `colData`, and includes the "ratio" assay.
}
\examples{
# Read example dataset
data <- readData(system.file(package = "mzQuality", "example.tsv"))

# Construct experiment
exp <- buildExperiment(
    data,
    rowIndex = "compound",
    colIndex = "aliquot",
    primaryAssay = "area",
    secondaryAssay = "area_is"
)

# Validate if the experiment is formatted correctly
validateExperiment(exp)
}
