% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{violinPlot}
\alias{violinPlot}
\title{Violin Plot}
\usage{
violinPlot(
  exp,
  assay = "ratio_corrected",
  batches = exp$batch,
  types = metadata(exp)$QC,
  doLog = FALSE,
  addMedian = TRUE,
  addConfidenceInterval = TRUE,
  withinTrend = FALSE
)
}
\value{
placeholder
}
\description{
A short description...
}
\details{
placeholder
}
