% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{writeExport}
\alias{writeExport}
\title{Write the new export file}
\usage{
writeExport(
  outFile,
  exp,
  types = exp$type,
  backgroundPercent = 40,
  cautionRSD = 15,
  nonReportableRSD = 30,
  digits = 3,
  selectedOnly = FALSE
)
}
\description{
Write the new export file
}
